fix: move version comparison logic to PluginValidation class
- Moved the `compareVer` method from `RPHValidator` to the `PluginValidation` class
- Updated the usage of `compareVer` in `RPHValidator` to use the new static method
- Added a check in `PluginValidation.updatePluginVersion` to skip updating the plugin version if the new version is lower than the current version
