+++ src/Functions/Processors/RPH/RPHValidator.ts
@@ -4,6 +4,7 @@ import { PluginType } from '../../../CustomTypes/Enums/PluginType';
 import { State } from '../../../CustomTypes/Enums/State';
 import { RPHLog } from '../../../CustomTypes/LogTypes/RPHLog';
 import { Plugin } from '../../../CustomTypes/MainTypes/Plugin';
+import { PluginValidation } from '../../Validations/Plugins';
 import { RPHAdvancedErrors } from './RPHAdvancedErrors';
 
 export class RPHValidator {
@@ -56,7 +57,7 @@ export class RPHValidator {
             break;
           }
           if (!logPlug.version || !cachePlug.version) continue;
-          const result = this.compareVer(logPlug.version, cachePlug.version);
+          const result = PluginValidation.compareVer(logPlug.version, cachePlug.version);
           if (result < 0) {
             logPlug.eaVersion = cachePlug.version;
             if (!log.outdated.some((x) => x.name === logPlug.name)) {
@@ -108,18 +109,5 @@ export class RPHValidator {
     log.validaterCompletedAt = new Date();
     log.elapsedTime = (new Date().getTime() - log.validaterStartedAt.getTime()).toString();
     return log;
-  }
-
-  private static compareVer(version1: string, version2: string): number {
-    const parts1 = version1.split('.').map(Number);
-    const parts2 = version2.split('.').map(Number);
-    const length = Math.max(parts1.length, parts2.length);
-    for (let i = 0; i < length; i++) {
-      const num1 = parts1[i] || 0;
-      const num2 = parts2[i] || 0;
-      if (num1 < num2) return -1;
-      if (num1 > num2) return 1;
-    }
-    return 0;
   }
 }
